Change Tabbar architecture to simple page elements
Turns out Tabbar is a big mess when it comes to routing since I did not
find a way to catch events when the active tab is changed. I switched to
using page elements with pushPage() and resetToPage() methods, this
implementation is very likely to be changed once v3 milestone is
reached.

Signed-off-by: Yidir <yader.dev@gmail.com>